fix under-counting of runs with shared cards
The previous commit fixed counting cards twice in runs like this:

  AH 2H 3H 4H

  AH 2H 3H

The second run shouldn't count, it's a proper subset of the first.

But these two should count:

  AH 2H 3H 4H

  AH 2H 3H 4S

Here, there cards that are the same, but they're two distinct runs.
Originally we only checked the low card, which made this fail but also
allowed the top case.  Then we fixed the top case, but it broke this
case.  I believe this fix is correct (finally).

It's important to remember that this is meant to work only because we
process subsets of the total hand in order of subset length.  If there
are two distinct subsets with length 4, they can both be runs of 4.  Any
set of length 3 that has a run only with cards in length 4 runs can't be
a new run.